Started adding the code to request to start a chat
Not finished yet, I need a way for the client to receive the request while they are in the menu or alternatively to have another option in the menu to see the message requests and accept/deny them from there. I have only done the client side and what I need will need to be on the server side.